+++ rust/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+## v0.4.0 (2024/05/25)
+* Add timeout options and oauth
+* Modify TiktokOauth new 
+
 ## v0.3.0 (2024/05/20)
 * Modify oauth_url parameters